CXR/model/.mdai/requirements.txt
numpy
scikit-image==0.19.3
torch==1.12.1
torchvision==0.13.1
pydicom==2.3.0
grad-cam
Pillow==9.5.0